move NixProfileListItem to nix.ProfileListItem
